decoder-py/requirements.txt
pyserial==3.5
autopep8
black
yapf